Refactor stats tests
