+++ mokapot/stats/statistics.py
@@ -112,83 +112,3 @@ class OnlineStatistics:
         return SummaryStatistics(
             self.n, self.min, self.max, self.sum, self.mean, self.var, self.sd
         )
-
-
-@typechecked
-class HistData:
-    bin_edges: np.ndarray[float]
-    counts: np.ndarray[int]
-
-    def __init__(self, bin_edges: np.ndarray[float], counts: np.ndarray[int]):
-        if len(bin_edges) != len(counts) + 1:
-            raise ValueError(
-                "`bin_edges` must have one more element than `counts` "
-                f"({len(bin_edges)=}, {len(counts)=})"
-            )
-
-        self.bin_edges = bin_edges
-        self.counts = counts
-
-    @property
-    def bin_centers(self) -> np.ndarray[float]:
-        return 0.5 * (self.bin_edges[:-1] + self.bin_edges[1:])
-
-    @property
-    def density(self) -> np.ndarray[float]:
-        dx = np.diff(self.bin_edges)
-        counts = self.counts.astype(float)
-        return counts / (dx * counts.sum())
-
-    @staticmethod
-    def bin_size_sturges(stats: OnlineStatistics) -> float:
-        return (stats.max - stats.min) / (np.log2(stats.n) + 1.0)
-
-    @staticmethod
-    def bin_size_scott(stats: OnlineStatistics) -> float:
-        factor = (24 * 24 * np.pi) ** (1.0 / 6.0)
-        return factor * stats.sd * stats.n ** (-1.0 / 3.0)
-
-    @staticmethod
-    def bin_size_terrell_scott(stats: OnlineStatistics) -> float:
-        num_bins = np.ceil((2.0 * stats.n) ** (1.0 / 3.0))
-        return (stats.max - stats.min) / num_bins
-
-    @staticmethod
-    def get_bin_edges(
-        stats: OnlineStatistics,
-        name="scott",
-        clip: tuple[int, int] | None = None,
-        extend: bool = False,
-    ):
-        if name == "scott":
-            bin_size = HistData.bin_size_scott(stats)
-        elif name == "terrell_scott":
-            bin_size = HistData.bin_size_terrell_scott(stats)
-        elif name == "sturges":
-            bin_size = HistData.bin_size_sturges(stats)
-        elif name == "auto":
-            bin_size = bin_size = HistData.bin_size_scott(stats)
-        else:
-            raise ValueError(f"Unrecognized binning algorithm name: {name}")
-
-        range = (stats.min, stats.max)
-        num_bins = int(np.ceil((range[1] - range[0]) / bin_size))
-        if clip is not None:
-            num_bins = np.clip(num_bins, *clip)
-
-        if extend:
-            bin_size = (range[1] - range[0]) / num_bins
-            num_bins += 1
-            if clip is not None:
-                num_bins = np.clip(num_bins, *clip)
-            range = (stats.min - 0.5 * bin_size, stats.max + 0.5 * bin_size)
-
-        bin_edges = np.histogram_bin_edges([], bins=num_bins, range=range)
-        return bin_edges
-
-
-def gaussian_iqr(mu: float, sigma: float) -> tuple[float, float]:
-    # Quartiles for the standard normal distribution are about +-0.67.
-    # Get the exact value with `scipy.stats.norm.isf(0.25)`.
-    alpha = 0.6744897501960817
-    return (mu - alpha * sigma, mu + alpha * sigma)
